24. DOM Scripting: Collection View
Note that the lines removed from index.html, landing.js as well as
removing utilities.js altogether are because the previous checkpoint’s
assignment should NOT have been merged. These lines should not be part
of the application from this point on. The code in landing.js reverts
back to the code from the previous checkpoint, prior to the assignment.